Safari-611.1.10/LayoutTests/media/media-source/media-source-stalled-holds-sleep-assertion-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(source.duration = loader.duration())
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(0)))
EVENT(update)
RUN(video.currentTime = video.buffered.end(0) - 0.1)
RUN(video.play())
EVENT(stalled)
EXPECTED (internals.elementIsBlockingDisplaySleep(video) == 'false') OK
END OF TEST
